requirements.txt
requests==2.32.5
prometheus-client==0.24.1
python-dotenv==1.2.1